add BaseEntity and AuditEntity
