+++ src/main/java/com/madikhan/app/model/AuditableEntity.java
@@ -1,0 +1,39 @@
+package com.madikhan.app.model;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.MappedSuperclass;
+import javax.persistence.PrePersist;
+import javax.persistence.PreUpdate;
+import java.io.Serializable;
+import java.time.LocalDateTime;
+
+@Data
+@EqualsAndHashCode(callSuper=true)
+@NoArgsConstructor
+@AllArgsConstructor
+@MappedSuperclass
+public abstract class AuditableEntity<T extends Serializable> extends BaseEntity<T> {
+
+    @JsonFormat(pattern = "yyyy-mm-dd HH::mm::ss")
+    @Column(updatable = false)
+    private LocalDateTime createdDate;
+
+    @JsonFormat(pattern = "yyyy-mm-dd HH::mm::ss")
+    private LocalDateTime updatedDate;
+
+    @PrePersist
+    protected void onCreate() {
+        this.createdDate = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    protected void onUpdate() {
+        this.updatedDate = LocalDateTime.now();
+    }
+}
