+++ src/main/java/com/madikhan/app/model/BaseEntity.java
@@ -1,11 +1,25 @@
 package com.madikhan.app.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
 import java.io.Serializable;
 
-public interface BaseEntity<T extends Serializable> {
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@MappedSuperclass
+public abstract class BaseEntity<T extends Serializable> {
 
-    void setId(T id);
-
-    T getId();
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private T id;
 
 }
