+++ src/main/java/com/madikhan/app/model/Birthday.java
@@ -1,28 +1,23 @@
 package com.madikhan.app.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
 public class Birthday {
 
     private LocalDate birthDate;
-
-    public Birthday() {
-    }
-
-    public Birthday(LocalDate birthDate) {
-        this.birthDate = birthDate;
-    }
 
     public long getAge() {
         return ChronoUnit.YEARS.between(birthDate, LocalDate.now());
     }
 
-    public LocalDate getBirthDate() {
-        return birthDate;
-    }
-
-    public void setBirthDate(LocalDate birthDate) {
-        this.birthDate = birthDate;
-    }
 }
