+++ src/main/java/com/madikhan/app/model/Comment.java
@@ -1,34 +1,26 @@
 package com.madikhan.app.model;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
-import javax.persistence.PrePersist;
 import javax.persistence.Table;
-import java.time.LocalDateTime;
 
 @Data
+@EqualsAndHashCode(callSuper=true)
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Entity
 @Table(name = "comment", schema = "social-network")
-public class Comment implements BaseEntity<Long> {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+public class Comment extends AuditableEntity<Long> {
 
     @ManyToOne(fetch = FetchType.EAGER)
     private Post post;
@@ -45,15 +37,4 @@ public class Comment implements BaseEntity<Long> {
     @OneToOne
     private Status status;
 
-    @JsonFormat(pattern = "yyyy-mm-dd HH::mm::ss")
-    @Column(updatable = false)
-    private LocalDateTime createdDate;
-
-    @Column
-    private LocalDateTime updatedDate;
-
-    @PrePersist
-    protected void onCreate() {
-        this.createdDate = LocalDateTime.now();
-    }
 }
